Deduplicate UnresolvableConfigurationResult
