+++ subprojects/diagnostics/src/main/java/org/gradle/api/tasks/diagnostics/internal/ConfigurationDetails.java
@@ -16,25 +16,13 @@
 
 package org.gradle.api.tasks.diagnostics.internal;
 
-import com.google.common.base.Joiner;
-import com.google.common.collect.Sets;
 import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.Dependency;
-import org.gradle.api.artifacts.DependencySet;
-import org.gradle.api.artifacts.ProjectDependency;
 import org.gradle.api.artifacts.result.ResolvedComponentResult;
 import org.gradle.api.provider.Provider;
-import org.gradle.api.tasks.diagnostics.internal.graph.nodes.AbstractRenderableDependency;
-import org.gradle.api.tasks.diagnostics.internal.graph.nodes.RenderableDependency;
-import org.gradle.api.tasks.diagnostics.internal.graph.nodes.UnresolvableConfigurationResult2;
+import org.gradle.api.tasks.diagnostics.internal.graph.nodes.UnresolvableConfigurationResult;
 import org.gradle.internal.deprecation.DeprecatableConfiguration;
 
 import javax.annotation.Nullable;
-import java.util.Collections;
-import java.util.Objects;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class ConfigurationDetails {
 
@@ -45,50 +33,13 @@ public class ConfigurationDetails {
             configuration.getDescription(),
             canBeResolved,
             canBeResolved ? configuration.getIncoming().getResolutionResult().getRootComponent() : null,
-            canBeResolved ? null : unresolvableResultOf(configuration)
+            canBeResolved ? null : UnresolvableConfigurationResult.of(configuration)
         );
     }
 
     private static boolean canBeResolved(Configuration configuration) {
         boolean isDeprecatedForResolving = ((DeprecatableConfiguration) configuration).getResolutionAlternatives() != null;
         return configuration.isCanBeResolved() && !isDeprecatedForResolving;
-    }
-
-    private static UnresolvableConfigurationResult2 unresolvableResultOf(Configuration configuration) {
-        return new UnresolvableConfigurationResult2(
-            configuration.getClass().getName() + configuration.hashCode(),
-            configuration.getName(),
-            unresolvableChildren(configuration)
-        );
-    }
-
-    private static Set<? extends RenderableDependency> unresolvableChildren(Configuration configuration) {
-        final DependencySet dependencies = configuration.getDependencies();
-        if (dependencies.isEmpty()) {
-            return Collections.emptySet();
-        }
-        Set<UnresolvableRenderableDependency> children = Sets.newLinkedHashSet();
-        for (final Dependency dependency : dependencies) {
-            children.add(new UnresolvableRenderableDependency(
-                dependency.getClass().getName() + dependency.hashCode(),
-                (dependency instanceof ProjectDependency)
-                    ? projectDependencyLabel((ProjectDependency) dependency)
-                    : moduleDependencyLabel(dependency)
-            ));
-        }
-        return children;
-    }
-
-    private static String projectDependencyLabel(ProjectDependency dependency) {
-        return "project " + dependency.getName();
-    }
-
-    private static String moduleDependencyLabel(Dependency dependency) {
-        return Joiner.on(":").join(
-            Stream.of(dependency.getGroup(), dependency.getName(), dependency.getVersion())
-                .filter(Objects::nonNull)
-                .collect(Collectors.toList())
-        );
     }
 
     private final String name;
@@ -102,14 +53,14 @@ public class ConfigurationDetails {
     private final Provider<ResolvedComponentResult> resolutionResultRoot;
 
     @Nullable
-    private final UnresolvableConfigurationResult2 unresolvableResult;
+    private final UnresolvableConfigurationResult unresolvableResult;
 
     private ConfigurationDetails(
         String name,
         @Nullable String description,
         boolean canBeResolved,
         @Nullable Provider<ResolvedComponentResult> resolutionResultRoot,
-        @Nullable UnresolvableConfigurationResult2 unresolvableResult
+        @Nullable UnresolvableConfigurationResult unresolvableResult
     ) {
         this.name = name;
         this.description = description;
@@ -137,28 +88,7 @@ public class ConfigurationDetails {
     }
 
     @Nullable
-    public UnresolvableConfigurationResult2 getUnresolvableResult() {
+    public UnresolvableConfigurationResult getUnresolvableResult() {
         return unresolvableResult;
-    }
-
-    private static final class UnresolvableRenderableDependency extends AbstractRenderableDependency {
-
-        private final Object id;
-        private final String name;
-
-        public UnresolvableRenderableDependency(Object id, String name) {
-            this.id = id;
-            this.name = name;
-        }
-
-        @Override
-        public Object getId() {
-            return id;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
     }
 }
