+++ subprojects/diagnostics/src/main/java/org/gradle/api/tasks/diagnostics/internal/graph/DependencyGraphsRenderer.java
@@ -19,7 +19,6 @@ package org.gradle.api.tasks.diagnostics.internal.graph;
 import com.google.common.collect.Sets;
 import org.gradle.api.tasks.diagnostics.internal.graph.nodes.RenderableDependency;
 import org.gradle.api.tasks.diagnostics.internal.graph.nodes.UnresolvableConfigurationResult;
-import org.gradle.api.tasks.diagnostics.internal.graph.nodes.UnresolvableConfigurationResult2;
 import org.gradle.internal.graph.GraphRenderer;
 import org.gradle.internal.logging.text.StyledTextOutput;
 
@@ -70,9 +69,6 @@ public class DependencyGraphsRenderer {
 
     private void renderRoot(final RenderableDependency root) {
         if (root instanceof UnresolvableConfigurationResult) {
-            legendRenderer.setHasUnresolvableConfigurations(true);
-        }
-        if (root instanceof UnresolvableConfigurationResult2) {
             legendRenderer.setHasUnresolvableConfigurations(true);
         }
         if (rootRenderer != NodeRenderer.NO_OP) {
