--- subprojects/diagnostics/src/main/java/org/gradle/api/tasks/diagnostics/internal/graph/nodes/UnresolvableConfigurationResult2.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.api.tasks.diagnostics.internal.graph.nodes;
-
-import java.util.Set;
-
-public class UnresolvableConfigurationResult2 extends AbstractRenderableDependency {
-
-    private final Object id;
-    private final String name;
-    private final Set<? extends RenderableDependency> children;
-
-    public UnresolvableConfigurationResult2(Object id, String name, Set<? extends RenderableDependency> children) {
-        this.id = id;
-        this.name = name;
-        this.children = children;
-    }
-
-    @Override
-    public Object getId() {
-        return id;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public Set<? extends RenderableDependency> getChildren() {
-        return children;
-    }
-}
